Rename `cli_context.py`.
