+++ tests_metricflow/cli/test_cli.py
@@ -19,7 +19,7 @@ from metricflow_semantics.test_helpers.example_project_configuration import (
     EXAMPLE_PROJECT_CONFIGURATION_YAML_CONFIG_FILE,
 )
 
-from dbt_metricflow.cli.cli_context import CLIContext
+from dbt_metricflow.cli.cli_configuration import CLIContext
 from dbt_metricflow.cli.main import (
     dimension_values,
     dimensions,
