+++ tests_metricflow/fixtures/cli_fixtures.py
@@ -12,7 +12,7 @@ from dbt_semantic_interfaces.protocols.semantic_manifest import SemanticManifest
 from metricflow_semantics.model.semantic_manifest_lookup import SemanticManifestLookup
 from typing_extensions import override
 
-from dbt_metricflow.cli.cli_context import CLIContext
+from dbt_metricflow.cli.cli_configuration import CLIContext
 from dbt_metricflow.cli.dbt_connectors.dbt_config_accessor import dbtArtifacts, dbtProjectMetadata
 from metricflow.engine.metricflow_engine import MetricFlowEngine
 from metricflow.protocols.sql_client import SqlClient
